refactor: Move security dependencies to common directory
