+++ src/adapter/common/bcrypt/bcrypt.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { BcryptAdapter } from '@adapter/security/bcrypt/bcrypt.adapter';
+import { BcryptAdapter } from '@adapter/common/bcrypt/bcrypt.adapter';
 
 @Module({
   providers: [BcryptAdapter],
