+++ src/adapter/in/web/service-proxy/admin.service-proxy.module.ts
@@ -7,16 +7,16 @@ import { ADMIN_ISSUE_INVITATION_USECASE } from '@application/port/in/admin/admin
 import { AdminIssueInvitationService } from '@application/service/admin/admin-issue-invitation.service';
 import { InvitationServiceProxyModule } from '@adapter/in/web/service-proxy/invitation.service-proxy.module';
 import { SIGN_UP_ADMIN_USECASE } from '@application/port/in/admin/sign-up-admin.usecase';
-import { BcryptAdapter } from '@adapter/security/bcrypt/bcrypt.adapter';
-import { BcryptPort } from '@application/port/security/bcrypt/bcrypt.port';
+import { BcryptAdapter } from '@adapter/common/bcrypt/bcrypt.adapter';
+import { BcryptPort } from '@application/port/common/bcrypt/bcrypt.port';
 import { SignUpAdminService } from '@application/service/admin/sign-up-admin.service';
-import { BcryptModule } from '@adapter/security/bcrypt/bcrypt.module';
+import { BcryptModule } from '@adapter/common/bcrypt/bcrypt.module';
 import { EnvironmentConfigService } from '@adapter/config/environment-config.service';
-import { JwtAdapter } from '@adapter/security/jwt/jwt.adapter';
+import { JwtAdapter } from '@adapter/common/jwt/jwt.adapter';
 import { ADMIN_TOKEN_USECASE } from '@application/port/in/auth/token/token.usecase';
 import { TokenService } from '@application/service/auth/token/token.service';
 import { EnvironmentConfigModule } from '@adapter/config/environment-config.module';
-import { JwtModule } from '@adapter/security/jwt/jwt.module';
+import { JwtModule } from '@adapter/common/jwt/jwt.module';
 import { SIGN_IN_ADMIN_USECASE } from '@application/port/in/admin/sign-in-admin.usecase';
 import { SignInAdminService } from '@application/service/admin/sign-in-admin.service';
 import { Admin } from '@domain/admin/admin';
