+++ src/adapter/in/web/service-proxy/auth.service-proxy.module.ts
@@ -11,14 +11,14 @@ import { VerificationAuthCodeService } from '@application/service/auth/verificat
 import { SIGN_UP_USECASE } from '@application/port/in/auth/sign-up.usecase';
 import { UserMongoRepository } from '@adapter/out/persistence/user/user.mongo.repository';
 import { SignUpService } from '@application/service/auth/sign-up.service';
-import { BcryptAdapter } from '@adapter/security/bcrypt/bcrypt.adapter';
-import { BcryptModule } from '@adapter/security/bcrypt/bcrypt.module';
+import { BcryptAdapter } from '@adapter/common/bcrypt/bcrypt.adapter';
+import { BcryptModule } from '@adapter/common/bcrypt/bcrypt.module';
 import { EnvironmentConfigService } from '@adapter/config/environment-config.service';
-import { JwtAdapter } from '@adapter/security/jwt/jwt.adapter';
+import { JwtAdapter } from '@adapter/common/jwt/jwt.adapter';
 import { TOKEN_USECASE } from '@application/port/in/auth/token/token.usecase';
 import { TokenService } from '@application/service/auth/token/token.service';
 import { EnvironmentConfigModule } from '@adapter/config/environment-config.module';
-import { JwtModule } from '@adapter/security/jwt/jwt.module';
+import { JwtModule } from '@adapter/common/jwt/jwt.module';
 import { SIGN_IN_USECASE } from '@application/port/in/auth/sign-in.usecase';
 import { SignInService } from '@application/service/auth/sign-in.service';
 import { LoggerAdapter } from '@adapter/common/logger/logger.adapter';
