+++ src/app.module.ts
@@ -4,8 +4,8 @@ import { DomainModule } from '@domain/domain.module';
 import { ApplicationModule } from '@application/application.module';
 import { AdapterModule } from '@adapter/adapter.module';
 import { EnvironmentConfigModule } from '@adapter/config/environment-config.module';
-import { BcryptModule } from '@adapter/security/bcrypt/bcrypt.module';
-import { JwtModule } from '@adapter/security/jwt/jwt.module';
+import { BcryptModule } from '@adapter/common/bcrypt/bcrypt.module';
+import { JwtModule } from '@adapter/common/jwt/jwt.module';
 import { JwtStrategy } from '@adapter/common/strategy/jwt.strategy';
 import { PassportModule } from '@nestjs/passport';
 import { LoggerModule } from '@adapter/common/logger/logger.module';
