+++ src/application/service/auth/token/token.service.ts
@@ -3,9 +3,9 @@ import { JwtConfig } from '@application/port/common/config/jwt.config';
 import {
   JwtPort,
   JwtServicePayload,
-} from '@application/port/security/jwt/jwt.port';
+} from '@application/port/common/jwt/jwt.port';
 import { AuthToken } from '@domain/user/auth-token';
-import { BcryptPort } from '@application/port/security/bcrypt/bcrypt.port';
+import { BcryptPort } from '@application/port/common/bcrypt/bcrypt.port';
 import { TokenRepository } from '@application/port/out/auth/token.repository';
 import { Accountable } from '@domain/auth/accountable';
